backend/pyproject.toml
[project]
name = "PolyRag-Backend"
version = "0.1.0"
description = "Backend service for PolyRAG - AI agent service built with LangGraph and FastAPI"
readme = "README.md"
authors = [{ name = "Joshua Carroll", email = "carroll.joshk@gmail.com" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

requires-python = ">=3.12"

dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "bcrypt>=4.3.0",
    "certifi==2025.1.31",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "colorama==0.4.6",
    "distro==1.9.0",
    "fastapi==0.115.12",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "jiter==0.9.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain-core==0.3.72",
    "langchain-openai==0.3.10",
    "langgraph==0.6.1",
    "langgraph-checkpoint==2.1.1",
    "langgraph-checkpoint-postgres>=2.0.23",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "langgraph-prebuilt==0.6.1",
    "langgraph-sdk==0.2.0",
    "llmsherpa>=0.1.4",
    "numexpr>=2.11.0",
    "openai==1.68.2",
    "orjson==3.10.16",
    "ormsgpack>=1.10.0",
    "packaging==24.2",
    "pip==25.0.1",
    "plotly>=6.0.1",
    "pandas>=2.2.0",
    "python-jose[cryptography]>=3.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic==2.10.6",
    "pydantic-core==2.27.2",
    "pydantic-settings>=2.10.1",
    "python-dotenv==1.1.0",
    "python-multipart>=0.0.20",
    "pyyaml==6.0.2",
    "regex==2024.11.6",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "sendgrid>=6.12.4",
    "sniffio==1.3.1",
    "sqlalchemy>=2.0.42",
    "starlette==0.46.1",
    "tenacity==9.0.0",
    "tiktoken==0.9.0",
    "tqdm==4.67.1",
    "typing-extensions==4.13.0",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "websockets==15.0.1",
    "xxhash==3.5.0",
    "zstandard==0.23.0",
    "python-jose==3.5.0",
    "pydantic[email]",
]

[dependency-groups]
dev = [
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pytest-env",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I", "U"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_default_fixture_loop_scope = "function"

[tool.pytest_env]
OPENAI_API_KEY = "sk-fake-openai-key"

[tool.mypy]
plugins = "pydantic.mypy"

[[tool.mypy.overrides]]
module = ["numexpr.*"]
follow_untyped_imports = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
